src/requirements.txt
accelerate==1.1.1
annotated-types==0.7.0
anyio==4.6.2.post1
Authlib==1.3.1
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
fsspec==2024.10.0
grpcio==1.68.0
grpcio-health-checking==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
openai==1.55.1
packaging==24.2
passlib==1.7.4
pillow==11.0.0
protobuf==5.28.3
psutil==6.1.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
simple-websocket==1.1.0
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.45.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.34.0
weaviate-client==4.9.4
Werkzeug==3.1.3
wsproto==1.2.0
